SLOBS frontend: soak first mac-virtualcam start event
* Commit c85d4fad7 renames virtualcam events into start & stop. Problem
is OBS frontend will always assume the user has started the virtualcam now.
Unfortunately now there is a modal popup that is
displayed if you attempt to close OBS because it thinks you started
virtualcam (when you didnt). This change at least gets rid of the modal
dialog when you run SLOBS backend